requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.3
aniso8601==9.0.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.2
ctransformers==0.2.22
dataclasses-json==0.5.14
docopt==0.6.2
dominate==2.8.0
faiss-cpu==1.7.4
filelock==3.12.2
Flask==2.3.2
flask-babel==3.1.0
Flask-Bootstrap==3.3.7.1
Flask-Migrate==4.0.4
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.5
Flask-Table==0.5.0
Flask-WTF==1.1.1
frozenlist==1.4.0
fsspec==2023.6.0
greenlet==2.0.2
huggingface-hub==0.16.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
langchain==0.0.268
langsmith==0.0.25
lit==16.0.6
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.5
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
Pillow==10.0.0
pipreqs==0.4.13
py-cpuinfo==9.0.0
pydantic==2.2.1
pydantic_core==2.6.1
PyPDF2==3.0.1
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rpds-py==0.9.2
safetensors==0.3.2
scikit-learn==1.3.0
scipy==1.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
SQLAlchemy==2.0.20
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.31.0
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
visitor==0.1.3
Werkzeug==2.3.7
WTForms==3.0.1
yarg==0.1.9
yarl==1.9.2
